-Fixed/Resolved issues: 21527, 21529;
git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS02]$/alphafs/Development;C111439

+++ AlphaFS/Filesystem/DirectoryInfo.cs
@@ -299,7 +299,11 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public override void Delete()
       {
+         if (!Exists)
+            throw new DirectoryNotFoundException(LongFullName);
+
          Directory.DeleteDirectoryInternal(EntryInfo, Transaction, null, false, false, true, false, null);
+         
          Reset();
       }
 
@@ -313,6 +317,9 @@ namespace Alphaleonis.Win32.Filesystem
       [SecurityCritical]
       public void Delete(bool recursive)
       {
+         if (!Exists)
+            throw new DirectoryNotFoundException(LongFullName);
+
          Directory.DeleteDirectoryInternal(EntryInfo, Transaction, null, recursive, false, !recursive, false, null);
          Reset();
       }
